+++ docs/topics/release-notes.md
@@ -8,6 +8,7 @@
 
 ### Master
 
+* Bugfix: Ensure read-only fields don't have model validation applied.
 * Bugfix: Fix hyperlinked fields in paginated results.
 
 ### 2.1.9